requirements.txt
pyroute2
ifcfg
pyyaml